requirements.txt
beautifulsoup4==4.3.2
bottle==0.12.7
requests==2.2.1
feedparser==5.1.3
python-mpd2==0.5.4
pafy
youtube-dl
